\name{data1}
\alias{data1}
\docType{data}
\title{
data1: Kaps and Lamberson(2009): page 252

}
\description{
The experiment compared three diets for pigs in a completely randomized design
}
\usage{data(data1)}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{Diet}}{a factor with levels \code{d1} \code{d2} \code{d3}}
    \item{\code{Gain}}{a numeric vector}
  }
}
\references{
KAPS, M. and LAMBERSON, W. R. Biostatistics for Animal Science: an introductory text. 2nd Edition. CABI Publishing, Wallingford, Oxfordshire, UK, 2009. 504p.
}
\examples{
data(data1)
summary(data1)
}

