\name{data11}
\alias{data11}
\docType{data}
\title{
data11: Pimentel Gomes and Garcia (2002): page 199
}
\description{
Incomplete block design
}
\usage{data(data11)}
\format{
  A data frame with 56 observations on the following 4 variables.
  \describe{
    \item{\code{treatments}}{a numeric vector}
    \item{\code{rep}}{a numeric vector}
    \item{\code{blocks}}{a numeric vector}
    \item{\code{yield}}{a numeric vector}
  }
}
\references{
PIMENTEL-GOMES, F. and GARCIA C.H. Estatistica aplicada a experimentos agronomicos e florestais: exposicao com exemplos e orientacoes para uso de aplicativos. Editora Fealq, v.11, 2002. 309p.
}
\examples{
data(data11)
summary(data11)
}

