\name{data12}
\alias{data12}
\docType{data}
\title{
data12: Pimentel Gomes and Garcia (2002): page 202
}
\description{
Incomplete block design
}
\usage{data(data12)}
\format{
  A data frame with 42 observations on the following 4 variables.
  \describe{
    \item{\code{treatments}}{a numeric vector}
    \item{\code{rep}}{a numeric vector}
    \item{\code{blocks}}{a numeric vector}
    \item{\code{yield}}{a numeric vector}
  }
}
\references{
PIMENTEL-GOMES, F. and GARCIA C.H. Estatistica aplicada a experimentos agronomicos e florestais: exposicao com exemplos e orientacoes para uso de aplicativos. Editora Fealq, v.11, 2002. 309p.
}
\examples{
data(data12)
summary(data12)
}

