\name{data16}
\alias{data16}
\docType{data}
\title{
data16: Sampaio (2010): page164
}
\description{
Switchback design
}
\usage{data(data16)}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{treatment}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{period}}{a numeric vector}
    \item{\code{animal}}{a numeric vector}
    \item{\code{gain}}{a numeric vector}
  }
}
\references{
SAMPAIO, I. B. M. Estatistica aplicada a experimentacao animal. 3nd Edition. Belo Horizonte: Editora FEPMVZ, Fundacao de Ensino e Pesquisa em Medicina Veterinaria e Zootecnia, 2010. 264p.
}
\examples{
data(data16)
summary(data16)
}

