\name{data17}
\alias{data17}
\docType{data}
\title{
data17: Sanders and Gaynor (1987)
}
\description{
Switchback design
}
\usage{data(data17)}
\format{
  A data frame with 36 observations on the following 5 variables.
  \describe{
    \item{\code{treatments}}{a numeric vector}
    \item{\code{blocks}}{a factor with levels \code{b1} \code{b2} \code{b3}}
    \item{\code{period}}{a numeric vector}
    \item{\code{animal}}{a numeric vector}
    \item{\code{gain}}{a numeric vector}
  }
}
\references{
SANDERS W.L. and GAYNOR, P.J. Analysis of switchback data using Statistical Analysis System, Inc. Software. Journal of Dairy Science, 70.2186-2191. 1987.
}
\examples{
data(data17)
summary(data17)
}

