\name{data18}
\alias{data18}
\docType{data}
\title{
data18: Ramalho et al. (2005): page 115
}
\description{
Repetition of experiments in block design
}
\usage{data(data18)}
\format{
  A data frame with 60 observations on the following 4 variables.
  \describe{
    \item{\code{treatments}}{a numeric vector}
    \item{\code{experiments}}{a numeric vector}
    \item{\code{blocks}}{a numeric vector}
    \item{\code{response}}{a numeric vector}
  }
}


\references{

RAMALHO, M. A. P.; FERREIRA, D. F. and OLIVEIRA, A. C. Experimentacao em Genetica e Melhoramento de Plantas. Editora UFLA, 2005, 322p.

}
\examples{
data(data18)
summary(data18)
}

