\name{data5}
\alias{data5}
\docType{data}
\title{
data5: Kaps and Lamberson (2009): page 361
}
\description{
Factorial in randomized design for testing two vitamins in feed of pigs
}
\usage{data(data5)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{Vitamin_1}}{a numeric vector}
    \item{\code{Vitamin_2}}{a numeric vector}
    \item{\code{Gains}}{a numeric vector}
  }
}
\references{
KAPS, M. and LAMBERSON, W. R. Biostatistics for Animal Science: an introductory text. 2nd Edition. CABI Publishing, Wallingford, Oxfordshire, UK, 2009. 504p.
}
\examples{
data(data5)
summary(data5)
}

