\name{data6}
\alias{data6}
\docType{data}
\title{
data6: Pimentel Gomes and Garcia (2002): page 127
}
\description{
Factorial in randomized block design
}
\usage{data(data6)}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{factor1}}{a numeric vector}
    \item{\code{factor2}}{a numeric vector}
    \item{\code{block}}{a numeric vector}
    \item{\code{yield}}{a numeric vector}
  }
}
\references{
PIMENTEL-GOMES, F. and GARCIA C.H. Estatistica aplicada a experimentos agronomicos e florestais: exposicao com exemplos e orientacoes para uso de aplicativos. Editora Fealq, v.11, 2002. 309p.
}
\examples{
data(data6)
summary(data6)
}

